--- test/controllers/article_list_controller_test.exs
@@ -1,13 +1,0 @@
-defmodule App.ArticleControllerTest do
-  use App.ConnCase
-
-  test "lists all entries on index", %{conn: conn} do
-    conn = get conn, article_list_path(conn, :index)
-    assert html_response(conn, 200) =~ "All Resources"
-  end
-
-  test "renders page not found when id is nonexistent", %{conn: conn} do
-    conn = get conn, article_list_path(conn, :show, "non-existent")
-    assert html_response(conn, 404) =~ "not found"
-  end
-end
