--- web/templates/article_list/index.html.eex
@@ -1,13 +1,0 @@
-<h2>All Resources</h2>
-
-<%= for resource <- @resources do %>
-  <%= if resource[:url] do %>
-    <div>
-      <%= link resource.heading, to: resource.url %>
-    </div>
-  <%= else %>
-    <div>
-      <%= link resource.heading, to: article_path(@conn, :show, resource.id) %>
-    </div>
-  <% end %>
-<% end %>
