--- web/templates/homepage/tag.html.eex
@@ -1,3 +1,0 @@
-<div>
-  <%= raw(@content) %>
-</div>
