--- web/views/article_list_view.ex
@@ -1,3 +1,0 @@
-defmodule App.ArticleListView do
-  use App.Web, :view
-end
